requirements.txt
streamlit
langchain
langchain-community
langchain_groq
chromadb
python-dotenv
pypdf
docx2txt
python-magic-bin
python-magic
tiktoken